adding limit to expenses and adding an alert dialog to warn user about their limit
