+++ lib/features/on_boarding/screen/on_boarding_screen.dart
@@ -1,5 +1,5 @@
 import 'package:basecode/components/intro_screen.dart';
-import 'package:basecode/services/auth/screen/log_in_screen.dart';
+import 'package:basecode/features/auth/screen/log_in_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
