+++ lib/features/user_profile/screen/edit_profile_screen.dart
@@ -2,9 +2,9 @@ import 'dart:io';
 
 import 'package:basecode/components/edit_profile_tile.dart';
 import 'package:basecode/model/user_model.dart';
-import 'package:basecode/services/auth/repository/auth_repository.dart';
-import 'package:basecode/services/user_profile/repository/storage_repository.dart';
-import 'package:basecode/services/user_profile/repository/user_profile_repository.dart';
+import 'package:basecode/features/auth/repository/auth_repository.dart';
+import 'package:basecode/features/user_profile/repository/storage_repository.dart';
+import 'package:basecode/features/user_profile/repository/user_profile_repository.dart';
 import 'package:basecode/utils/pick_image.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -13,11 +13,14 @@ class EditProfileScreen extends StatefulWidget {
   final String name;
   final String phoneNumber;
   final double budget;
-  const EditProfileScreen(
-      {super.key,
-      required this.name,
-      required this.phoneNumber,
-      required this.budget});
+  final double limit;
+  const EditProfileScreen({
+    super.key,
+    required this.name,
+    required this.phoneNumber,
+    required this.budget,
+    required this.limit,
+  });
 
   @override
   State<EditProfileScreen> createState() => _EditProfileScreenState();
@@ -27,9 +30,11 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
   final TextEditingController nameController = TextEditingController();
   final TextEditingController phoneController = TextEditingController();
   final TextEditingController budgetController = TextEditingController();
+  final TextEditingController limitController = TextEditingController();
   File? profileFile;
 
-  void saveChanges(UserModel user, File? profileFile, BuildContext context) async {
+  void saveChanges(
+      UserModel user, File? profileFile, BuildContext context) async {
     if (profileFile != null) {
       final res = await context.read<StorageRepository>().storeFile(
             path: 'user/profile',
@@ -42,6 +47,7 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
             name: nameController.text,
             phoneNumber: phoneController.text,
             budget: double.parse(budgetController.text),
+            limit: double.parse(limitController.text),
             avatar: res,
           ),
           context);
@@ -71,6 +77,7 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
     nameController.text = widget.name;
     phoneController.text = widget.phoneNumber;
     budgetController.text = widget.budget.toString();
+    limitController.text = widget.limit.toString();
   }
 
   @override
@@ -94,7 +101,10 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
               ),
               actions: [
                 TextButton(
-                  onPressed: () => saveChanges(snapshot.data!,profileFile, context),
+                  onPressed: () {
+                    saveChanges(snapshot.data!, profileFile, context);
+                    Navigator.of(context).pop();
+                  },
                   child: Text(
                     "Save",
                     style: TextStyle(
@@ -138,6 +148,11 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
                     EditProfileTile(
                       leading: "Budget",
                       contoller: budgetController,
+                    ),
+                    const SizedBox(height: 20),
+                    EditProfileTile(
+                      leading: "limit",
+                      contoller: limitController,
                     ),
                   ],
                 ),
