+++ pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
   uuid: ^4.4.0
   file_picker: ^8.0.6
   firebase_storage: ^12.1.1
+  badges: ^3.1.2
 
 dev_dependencies:
   flutter_test:
